Finished writing Rust structs for all tables

+++ web/backend/migrations/00000000000004_create_archivables_table/up.sql
@@ -1,6 +1,6 @@
 -- SQL defining the archivables table, containing all archivable content.
 CREATE TABLE archivables (
-    id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
     archived_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     archived_by INTEGER NOT NULL REFERENCES users(id)
 );
