+++ web/backend/migrations/00000000000020_create_item_units_table/up.sql
@@ -4,8 +4,8 @@
 -- measure items. Some items may be measured reasonably in different units, and this table
 -- allows for the definition of the units of measure that are used to measure items.
 CREATE TABLE item_units (
-    id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
-    item_id INTEGER NOT NULL REFERENCES items(id) ON DELETE CASCADE,
-    unit_id INTEGER NOT NULL REFERENCES units(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+    item_id BIGINT NOT NULL REFERENCES items(id) ON DELETE CASCADE,
+    unit_id BIGINT NOT NULL REFERENCES units(id) ON DELETE CASCADE,
     UNIQUE (item_id, unit_id)
 );
