+++ web/backend/migrations/00000000000023_create_item_continuous_quantities_table/up.sql
@@ -3,11 +3,11 @@
 -- The weight of an item may change over time, and be measured using different scales and by
 -- different people.
 CREATE TABLE item_continuous_quantities (
-    id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
-    item_id INTEGER REFERENCES items(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+    item_id BIGINT REFERENCES items(id) ON DELETE CASCADE,
     weight DECIMAL(10, 2) NOT NULL,
-    unit_id INTEGER REFERENCES units(id) ON DELETE CASCADE,
-    sensor_id INTEGER REFERENCES items(id) ON DELETE SET NULL,
+    unit_id BIGINT REFERENCES units(id) ON DELETE CASCADE,
+    sensor_id BIGINT REFERENCES items(id) ON DELETE SET NULL,
     measured_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     measured_by INTEGER REFERENCES users(id) ON DELETE SET NULL,
     UNIQUE (item_id, unit_id),
