+++ web/backend/migrations/00000000000025_create_spectra_collections_table/up.sql
@@ -1,5 +1,5 @@
 -- Your SQL goes here
 CREATE TABLE spectra_collection (
-    id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
-    sample_id INTEGER REFERENCES samples(id) ON DELETE CASCADE NOT NULL
+    id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+    sample_id BIGINT REFERENCES samples(id) ON DELETE CASCADE NOT NULL
 );
