+++ web/backend/migrations/00000000000032_create_organization_users_table/up.sql
@@ -7,9 +7,9 @@
 -- Since an administrator needs to add a user to an organization, the organization_users table
 -- also contains a column to specify which administrator added the user to the organization.
 CREATE TABLE organization_users (
-    id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+    id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
     user_id INTEGER NOT NULL REFERENCES users (id) ON DELETE CASCADE,
-    organization_id INTEGER NOT NULL REFERENCES organizations (id) ON DELETE CASCADE,
-    role_id INTEGER NOT NULL REFERENCES organization_user_roles (id) ON DELETE SET NULL,
+    organization_id BIGINT NOT NULL REFERENCES organizations (id) ON DELETE CASCADE,
+    role_id BIGINT NOT NULL REFERENCES organization_user_roles (id) ON DELETE SET NULL,
     UNIQUE (user_id, organization_id, role_id)
 );
