+++ web/backend/migrations/00000000000036_create_organization_locations_table/up.sql
@@ -2,8 +2,8 @@
 -- An organization may be present in one or more locations.
 -- An organization may change locations over time.
 CREATE TABLE organization_locations (
-  id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
-  organization_id INTEGER REFERENCES organizations(id) ON DELETE CASCADE,
-  location_id INTEGER REFERENCES locations(id) ON DELETE CASCADE,
-  previous_location_id INTEGER REFERENCES organization_locations(id) ON DELETE SET NULL
+  id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+  organization_id BIGINT REFERENCES organizations(id) ON DELETE CASCADE,
+  location_id BIGINT REFERENCES locations(id) ON DELETE CASCADE,
+  previous_location_id BIGINT REFERENCES organization_locations(id) ON DELETE SET NULL
 );
