+++ web/backend/migrations/00000000000040_create_manifactured_item_categories_table/up.sql
@@ -3,9 +3,9 @@
 -- has a cost, its currency, and a cost per day for renting and the current of the renting
 -- cost.
 CREATE TABLE manufactured_item_categories (
-  id INTEGER PRIMARY KEY REFERENCES item_categories(id) ON DELETE CASCADE,
+  id BIGINT PRIMARY KEY REFERENCES item_categories(id) ON DELETE CASCADE,
   cost DECIMAL(10, 2) NOT NULL,
   cost_per_day DECIMAL(10, 2) NOT NULL,
   currency VARCHAR(3) NOT NULL,
-  manifacturer_id INTEGER REFERENCES organizations(id)
+  manifacturer_id BIGINT REFERENCES organizations(id)
 );
