+++ web/backend/migrations/00000000000042_create_item_category_relationships_table/up.sql
@@ -9,7 +9,7 @@
 -- columns are timestamps that are automatically set to the current time when a new
 -- row is created or updated.
 CREATE TABLE item_category_relationships (
-  id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+  id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
   parent_id INTEGER NOT NULL REFERENCES item_categories(id),
   child_id INTEGER NOT NULL REFERENCES item_categories(id),
   UNIQUE (parent_id, child_id)
