+++ web/backend/migrations/00000000000047_create_procedure_continuous_requirements_table/up.sql
@@ -2,11 +2,11 @@
 -- A procedure may require one or more items to be performed. This table defines the
 -- requirements of a procedure in terms of the number of items of a given category.
 CREATE TABLE procedure_continuous_requirements (
-  id INTEGER PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
-  procedure_id INTEGER NOT NULL REFERENCES procedures(id) ON DELETE CASCADE,
-  item_category_id INTEGER NOT NULL REFERENCES item_categories(id) ON DELETE CASCADE,
+  id BIGINT PRIMARY KEY REFERENCES editables(id) ON DELETE CASCADE,
+  procedure_id BIGINT NOT NULL REFERENCES procedures(id) ON DELETE CASCADE,
+  item_category_id BIGINT NOT NULL REFERENCES item_categories(id) ON DELETE CASCADE,
   quantity FLOAT NOT NULL,
-  unit_id INTEGER REFERENCES units(id) ON DELETE CASCADE,
+  unit_id BIGINT REFERENCES units(id) ON DELETE CASCADE,
   UNIQUE (procedure_id, item_category_id),
   FOREIGN KEY (item_category_id, unit_id) REFERENCES item_category_units(item_category_id, unit_id) ON DELETE CASCADE,
   FOREIGN KEY (unit_id) REFERENCES continuous_units(id) ON DELETE CASCADE
