+++ web/backend/migrations/00000000000051_create_sample_taxa_table/up.sql
@@ -4,7 +4,7 @@
 -- A sample can be associated to no or more taxa, and a taxon can be found in multiple samples.
 --
 CREATE TABLE sample_taxa (
-  sample_id INTEGER NOT NULL REFERENCES samples(id) ON DELETE CASCADE,
-  taxon_id INTEGER NOT NULL REFERENCES taxa(id) ON DELETE CASCADE,
+  sample_id BIGINT NOT NULL REFERENCES samples(id) ON DELETE CASCADE,
+  taxon_id BIGINT NOT NULL REFERENCES taxa(id) ON DELETE CASCADE,
   PRIMARY KEY (sample_id, taxon_id)
 );
